Read me-v3
the access port is the main part that i need in order for my 2008 WRX to run the way it is supposed to. it is in my greatest understanding the heart of the car. it is a full tuner race car in which needs specific modifications done. it is necessary that I understand what I am doing and this is a great way to put a collective amount of ideas in which I  can distribute what is in my mind much easier here.